refactor: Move to utilities.
