+++ src/scripts/utilities/copyCommonFiles.ts
