+++ src/scripts/utilities/writeEnvFile.ts
